backend/ai-recommendation/requirements.txt
flask
flask-cors
python-dotenv
supabase
pyjwt
openai
python-dateutil
